+++ changes.md
@@ -1,5 +1,7 @@
 # Guara
 
+ - Reworking how background services start
+
 ## Release v1.0.1
 on: 21/Feb/2025